LayoutTests/platform/glib/imported/w3c/web-platform-tests/css/css-text/white-space/white-space-collapse-002-expected.txt


PASS RLO 1
FAIL RLO 2 assert_equals: expected 104 but got 111
FAIL RLO 3 assert_equals: expected 104 but got 111
FAIL RLO 4 assert_equals: expected 104 but got 111
FAIL RLO 5 assert_equals: expected 104 but got 111
PASS RLE 1
FAIL RLE 2 assert_equals: expected 101 but got 108
FAIL RLE 3 assert_equals: expected 101 but got 108
FAIL RLE 4 assert_equals: expected 101 but got 108
FAIL RLE 5 assert_equals: expected 101 but got 108
PASS RLI 1
FAIL RLI 2 assert_equals: expected 92 but got 99
FAIL RLI 3 assert_equals: expected 92 but got 99
FAIL RLI 4 assert_equals: expected 92 but got 99
FAIL RLI 5 assert_equals: expected 92 but got 99
PASS RLM 1
FAIL RLM 2 assert_equals: expected 110 but got 117
FAIL RLM 3 assert_equals: expected 110 but got 117
FAIL RLM 4 assert_equals: expected 110 but got 117
FAIL RLM 5 assert_equals: expected 110 but got 117
RLO‮ level‬here
RLOlevel here
RLO ‮ level‬
RLO level
RLO ‮ level‬
RLO level
RLO ‮ level‬
RLO level
RLO ‮ level‬
RLO level
RLE‫ level‬here
RLElevel here
RLE ‫ level‬
RLE level
RLE ‫ level‬
RLE level
RLE ‫ level‬
RLE level
RLE ‫ level‬
RLE level
RLI⁧ level⁩here
RLIlevel here
RLI ⁧ level⁩
RLI level
RLI ⁧ level⁩
RLI level
RLI ⁧ level⁩
RLI level
RLI ⁧ level⁩
RLI level
RLM‏ mark
RLM mark
RLM ‏ mark
RLM mark
RLM ‏ mark
RLM mark
RLM ‏ mark
RLM mark
RLM ‏ mark
RLM mark
